// frontend/src/api/models/dtoAuditLogResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { DtoAuditLogResponseNewValue } from './dtoAuditLogResponseNewValue'
import type { DtoAuditLogResponseOldValue } from './dtoAuditLogResponseOldValue'

export interface DtoAuditLogResponse {
  action?: string
  created_at?: string
  flag_key?: string
  id?: string
  ip_address?: string
  new_value?: DtoAuditLogResponseNewValue
  old_value?: DtoAuditLogResponseOldValue
  user_agent?: string
  user_id?: string
}